# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReceiptLine": {
        "description": "Una riga di una ricezione merce.\n\nRisoluzione SKU (in ordine di precedenza):\n1. ``sku``  — codice SKU esatto (case-sensitive).\n2. ``ean``  — EAN-12/13; il server risolve l'EAN → SKU prima della validazione.\n3. Se entrambi mancano → errore 400 su quella riga.\nSe entrambi presenti, ``sku`` ha la precedenza; ``ean`` viene ignorato.",
        "properties": {
          "ean": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "EAN-12 o EAN-13. Alternativa a sku (lookup server-side).",
            "title": "Ean"
          },
          "expiry_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Data di scadenza del lotto. Obbligatoria se lo SKU ha has_expiry_label=true (etichetta scadenza manuale).",
            "title": "Expiry Date"
          },
          "note": {
            "default": "",
            "maxLength": 200,
            "title": "Note",
            "type": "string"
          },
          "qty_received": {
            "description": "Quantità ricevuta. 0 = nessun articolo ricevuto (riga accettata ma nessun evento RECEIPT scritto nel ledger).",
            "minimum": 0.0,
            "title": "Qty Received",
            "type": "integer"
          },
          "sku": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Codice SKU (case-sensitive). Obbligatorio se ean non è fornito.",
            "title": "Sku"
          }
        },
        "required": [
          "qty_received"
        ],
        "title": "ReceiptLine",
        "type": "object"
      },
      "ReceiptLineResult": {
        "properties": {
          "ean": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ean"
          },
          "expiry_date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expiry Date"
          },
          "line_index": {
            "title": "Line Index",
            "type": "integer"
          },
          "qty_received": {
            "title": "Qty Received",
            "type": "integer"
          },
          "sku": {
            "title": "Sku",
            "type": "string"
          },
          "status": {
            "enum": [
              "ok",
              "already_received",
              "skipped"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "line_index",
          "sku",
          "qty_received",
          "status"
        ],
        "title": "ReceiptLineResult",
        "type": "object"
      },
      "ReceiptsCloseRequest": {
        "properties": {
          "client_receipt_id": {
            "anyOf": [
              {
                "maxLength": 128,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "UUID opzionale del client per idempotenza forte. Se la stessa stringa arriva una seconda volta il server risponde already_posted=true senza riscrivere il ledger.",
            "examples": [
              "550e8400-e29b-41d4-a716-446655440000"
            ],
            "title": "Client Receipt Id"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/ReceiptLine"
            },
            "minItems": 1,
            "title": "Lines",
            "type": "array"
          },
          "receipt_date": {
            "format": "date",
            "title": "Receipt Date",
            "type": "string"
          },
          "receipt_id": {
            "description": "Chiave di idempotenza legacy (date+supplier+ref). Se già presente nei receiving_logs → 200 already_posted.",
            "maxLength": 100,
            "title": "Receipt Id",
            "type": "string"
          }
        },
        "required": [
          "receipt_id",
          "receipt_date",
          "lines"
        ],
        "title": "ReceiptsCloseRequest",
        "type": "object"
      },
      "ReceiptsCloseResponse": {
        "properties": {
          "already_posted": {
            "default": false,
            "title": "Already Posted",
            "type": "boolean"
          },
          "client_receipt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Receipt Id"
          },
          "lines": {
            "items": {
              "$ref": "#/components/schemas/ReceiptLineResult"
            },
            "title": "Lines",
            "type": "array"
          },
          "receipt_date": {
            "format": "date",
            "title": "Receipt Date",
            "type": "string"
          },
          "receipt_id": {
            "title": "Receipt Id",
            "type": "string"
          }
        },
        "required": [
          "receipt_id",
          "receipt_date",
          "lines"
        ],
        "title": "ReceiptsCloseResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "REST backend for **desktop-order-system**: stock queries, EAN lookup, exception logging, and receipt closure.\n\nAll write endpoints require a `Bearer <DOS_API_TOKEN>` header.\nSee [`docs/api_contract.md`](https://github.com/sasu91/desktop-order-system/blob/main/docs/api_contract.md) for the full contract.",
    "title": "desktop-order-system API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/receipts/close": {
      "post": {
        "description": "Registra gli eventi RECEIPT nel ledger per ogni riga ricevuta.\n\n**Atomicità**: se anche una sola riga fallisce validazione, **nessuna**\nriga viene scritta e la risposta 400 elenca tutti gli errori con indice\nriga, campo e motivo.\n\n**Idempotenza**:\n- ``client_receipt_id`` (UUID) → lookup in ``api_idempotency_keys``.\n  Duplicato → 200 ``already_posted=true``, ledger invariato.\n- ``receipt_id`` (legacy) → lookup in receiving_logs.\n  Già presente → 200 ``already_posted=true``.\n\n**Risoluzione SKU**:\n- ``sku`` esplicito → ricerca diretta.\n- ``ean`` → validazione formato + lookup catalogo → sku.\n- Se lo SKU ha ``has_expiry_label=true`` → ``expiry_date`` obbligatoria.",
        "operationId": "close_receipt_api_v1_receipts_close_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReceiptsCloseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptsCloseResponse"
                }
              }
            },
            "description": "Già processato (already_posted=true)."
          },
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReceiptsCloseResponse"
                }
              }
            },
            "description": "Ricezione registrata con successo."
          },
          "400": {
            "description": "Errori di validazione per riga — nessuna riga scritta."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Chiudi ricezione merce",
        "tags": [
          "receipts"
        ]
      }
    },
...
  }
}